Registration and Onboarding (and some assets)
The onboarding design has been redesigned, and a couple of lines have been changed to ensure proper asset uploads. Registration now includes the network synchronization step for the ecosystem. This last step needs some revisions to manage a synchronized network as a minimum, and to add an "undo" option when the user finishes editing/adding networks.